refactor: split init.lua into focused modules
- config.lua  : shared state (cfg, effective_repos, languages)
- util.lua    : path helpers (ppath/qpath/ext) and async utilities
- installer.lua : install/remove/update logic and repo queries
- ui.lua      : window rendering, keymaps, status icons
- init.lua    : setup() entry point only (394 -> 54 lines)

No behavior changes.
